move proxy check
