+++ qa/qa/checks/__init__.py
@@ -3,7 +3,7 @@
 from qa.checks.base import BaseCheck  # noqa
 from qa.checks.metadata.abstract import AbstractIsValid  # noqa
 from qa.checks.metadata.acm_class import AcmClassIsValid  # noqa
-from qa.checks.metadata.authors import AuthorsAreValid, AuthorsContainsSubmitterName  # noqa
+from qa.checks.metadata.authors import AuthorsAreValid  # noqa
 from qa.checks.metadata.comments import CommentsAreValid  # noqa
 from qa.checks.metadata.doi import DoiIsValid  # noqa
 from qa.checks.metadata.journal_ref import JournalRefIsValid  # noqa
@@ -13,6 +13,7 @@ from qa.checks.metadata.title import TitleIsValid  # noqa
 
 from qa.checks.submission.files import DoesNotExceedTheFileSizeLimit, FileTypeDoesNotRequireReview  # noqa
 from qa.checks.submission.type import IsNotAWithdrawal  # noqa
+from qa.checks.submission.user import AuthorsContainsSubmitterName  # noqa
 
 from qa.checks.fulltext.extraction import TextExtractionSuccessful  # noqa
 from qa.checks.fulltext.structure import FulltextNotTooShort  # noqa
